move keymap handling
