+++ napari/components/viewer_model.py
@@ -16,7 +16,7 @@ from ..utils import config
 from ..utils._register import create_func as create_add_method
 from ..utils.colormaps import ensure_colormap
 from ..utils.events import EmitterGroup, Event, disconnect_events
-from ..utils.key_bindings import KeymapHandler, KeymapProvider
+from ..utils.key_bindings import KeymapProvider
 from ..utils.misc import is_sequence
 from ..utils.theme import palettes
 from ._viewer_mouse_bindings import dims_scroll
@@ -29,7 +29,7 @@ from .layerlist import LayerList
 from .scale_bar import ScaleBar
 
 
-class ViewerModel(KeymapHandler, KeymapProvider):
+class ViewerModel(KeymapProvider):
     """Viewer containing the rendered scene, layers, and controlling elements
     including dimension sliders, and control bars for color limits.
 
@@ -110,8 +110,6 @@ class ViewerModel(KeymapHandler, KeymapProvider):
         self.layers.events.removed.connect(self._on_remove_layer)
         self.layers.events.reordered.connect(self._on_grid_change)
         self.layers.events.reordered.connect(self._on_layers_change)
-
-        self.keymap_providers = [self]
 
         # Hold callbacks for when mouse moves with nothing pressed
         self.mouse_move_callbacks = []
@@ -282,14 +280,7 @@ class ViewerModel(KeymapHandler, KeymapProvider):
         if active_layer == self.active_layer:
             return
 
-        if self._active_layer is not None:
-            self.keymap_providers.remove(self._active_layer)
-
         self._active_layer = active_layer
-
-        if active_layer is not None:
-            self.keymap_providers.insert(0, active_layer)
-
         self.events.active_layer(item=self._active_layer)
 
     @property
